make timer a separate function file
